Fix bug where you could not have just an f operator without a leading digit. Defaults to 1 if missing number of fudge dice. Refactored fudge dice tests into their own unit test class.
